Add graphrag-common package and move base factory to common.
