+++ packages/graphrag-common/graphrag_common/factory/__init__.py
@@ -1,8 +1,8 @@
 # Copyright (c) 2024 Microsoft Corporation.
 # Licensed under the MIT License
 
-"""The GraphRAG Factory package."""
+"""The GraphRAG factory module."""
 
-from graphrag_factory.factory import Factory
+from graphrag_common.factory.factory import Factory
 
 __all__ = ["Factory"]
