+++ packages/graphrag/graphrag/language_model/providers/litellm/services/rate_limiter/rate_limiter_factory.py
@@ -3,7 +3,7 @@
 
 """LiteLLM Rate Limiter Factory."""
 
-from graphrag_factory import Factory
+from graphrag_common.factory import Factory
 
 from graphrag.language_model.providers.litellm.services.rate_limiter.rate_limiter import (
     RateLimiter,
