+++ packages/graphrag/pyproject.toml
@@ -40,7 +40,7 @@ dependencies = [
     "azure-storage-blob>=12.24.0",
     "devtools>=0.12.2",
     "environs>=11.0.0",
-    "graphrag-factory==2.7.0",
+    "graphrag-common==2.7.0",
     "graspologic-native>=1.2.5",
     "json-repair>=0.30.3",
     "lancedb>=0.17.0",
