+++ pyproject.toml
@@ -53,7 +53,7 @@ package = false
 members = ["packages/*"]
 
 [tool.uv.sources]
-graphrag-factory = { workspace = true }
+graphrag-common = { workspace = true }
 
 # Keep poethepoet for task management to minimize changes
 [tool.poe.tasks]
@@ -68,7 +68,7 @@ _semversioner_release = "semversioner release"
 _semversioner_changelog = "semversioner changelog > CHANGELOG.md"
 # Add more update toml tasks as packages are added
 _semversioner_update_graphrag_toml_version = "update-toml update --file packages/graphrag/pyproject.toml --path project.version --value $(uv run semversioner current-version)"
-_semversioner_update_graphrag_factory_toml_version = "update-toml update --file packages/graphrag-factory/pyproject.toml --path project.version --value $(uv run semversioner current-version)"
+_semversioner_update_graphrag_common_toml_version = "update-toml update --file packages/graphrag-common/pyproject.toml --path project.version --value $(uv run semversioner current-version)"
 _semversioner_update_workspace_dependency_versions = "python -m scripts.update_workspace_dependency_versions"
 semversioner_add = "semversioner add-change"
 coverage_report = 'coverage report --omit "**/tests/**" --show-missing'
@@ -102,7 +102,7 @@ sequence = [
     '_semversioner_changelog',
     # Add more update toml tasks as packages are added
     '_semversioner_update_graphrag_toml_version',
-    '_semversioner_update_graphrag_factory_toml_version',
+    '_semversioner_update_graphrag_common_toml_version',
     '_semversioner_update_workspace_dependency_versions',
     '_sync',
 ]
@@ -217,7 +217,7 @@ convention = "numpy"
 
 # https://github.com/microsoft/pyright/blob/9f81564a4685ff5c55edd3959f9b39030f590b2f/docs/configuration.md#sample-pyprojecttoml-file
 [tool.pyright]
-include = ["packages/graphrag/graphrag", "packages/graphrag-factory/graphrag_factory", "tests"]
+include = ["packages/graphrag/graphrag", "packages/graphrag-common/graphrag_common", "tests"]
 exclude = ["**/node_modules", "**/__pycache__"]
 
 [tool.pytest.ini_options]
