+++ tests/unit/graphrag_factory/test_factory.py
@@ -5,7 +5,7 @@
 
 from abc import ABC, abstractmethod
 
-from graphrag_factory import Factory
+from graphrag_common.factory import Factory
 
 
 class TestABC(ABC):
